完善了其他细节，优化了处理效率，使用多线程来提高响应速度，又修复了user模块的部分bug。

+++ TheHomeOfDinners/apps/restaurant/views.py
@@ -1,7 +1,7 @@
-import json
 import os
+from threading import Thread
 
-from django.db.models import Q, F, Avg, Count
+from django.db.models import Q, F, Count
 from rest_framework import status
 # Create your views here.
 from rest_framework.decorators import action
@@ -11,9 +11,11 @@ from rest_framework.views import APIView
 from rest_framework.viewsets import ModelViewSet
 
 import parameters
+from TheHomeOfDinners.AIModule.analyze import analyze
+from TheHomeOfDinners.settings import BASE_DIR
 from restaurant.serializers import RestaurantSerializer, CollectionSerializer, MenuSerializer, ReviewSerializer
 from .models import Restaurant, Tag, Collection, Menu, Review
-from .utils import MyPageNumberPagination
+from .utils import MyPageNumberPagination, generateWordCloud
 
 
 class RestaurantModelViewSet(ModelViewSet):
@@ -21,13 +23,28 @@ class RestaurantModelViewSet(ModelViewSet):
     serializer_class = RestaurantSerializer
     pagination_class = MyPageNumberPagination
 
+    def list(self, request, *args, **kwargs):
+        queryset = self.filter_queryset(self.get_queryset())
+        # 根据参数名进行查询
+        if 'res_name' in request.GET:
+            queryset = queryset.filter(res_name__contains=request.GET['res_name'])
+        page = self.paginate_queryset(queryset)
+        if page is not None:
+            serializer = self.get_serializer(page, many=True)
+            return self.get_paginated_response(serializer.data)
+
+        serializer = self.get_serializer(queryset, many=True)
+        return Response(serializer.data)
+
     def update(self, request, *args, **kwargs):
+        """更新餐馆信息"""
         partial = kwargs.pop('partial', False)
         instance = self.get_object()
 
         try:
             # 删除之前的图片
-            os.remove(instance.picture.path)
+            if request.data['picture']:
+                os.remove(instance.picture.path)
         except Exception:
             pass
 
@@ -44,6 +61,7 @@ class RestaurantModelViewSet(ModelViewSet):
 
     @action(methods=['get'], detail=True)
     def reviews_count(self, request, pk):
+        """获取餐馆的评分分布情况"""
         counts = Review.objects.filter(restaurant=pk) \
             .values('score') \
             .annotate(count=Count('score')) \
@@ -53,13 +71,44 @@ class RestaurantModelViewSet(ModelViewSet):
             res[count[0]] = count[1]
         return Response(res)
 
+    @action(methods=['get'], detail=True)
+    def user_collections(self, request, pk):
+        """根据用户id获取收藏的餐馆列表"""
+        restaurants = Restaurant.objects.filter(restaurant_collection__user_id=pk)
+        restaurants = self.paginate_queryset(restaurants)
+        serializer = self.get_serializer(restaurants, many=True)
+        return self.get_paginated_response(serializer.data)
+
+    @action(methods=['get'], detail=True)
+    def pos_or_nav_reviews_count(self, request, pk):
+        reviews = Review.objects.filter(restaurant=pk)
+        pos_count = reviews.filter(analyze_result=0).count()
+        nav_count = reviews.filter(analyze_result=1).count()
+        return Response({'pos': pos_count, 'nav': nav_count})
+
+    @action(methods=['get'], detail=True)
+    def wordCloud(self, request, pk):
+        """获取词云图片，传入餐馆id"""
+        if os.path.exists(os.path.join(BASE_DIR, 'media', 'wordClouds', pk + ".png")):
+            return Response(os.path.join('pictures', 'wordClouds', pk + '.png'))
+        reviews = Review.objects.filter(restaurant=pk).values_list('text')
+        res = []
+        for review in reviews:
+            res.append(review[0])
+            print(''.join(res))
+        return Response(generateWordCloud(''.join(res), pk))
+
 
 class TagRestaurantDetailView(GenericAPIView):
+    queryset = Restaurant.objects.all()
     serializer_class = RestaurantSerializer
     pagination_class = MyPageNumberPagination
 
     def get(self, request, param):
-        """根据标签查询餐馆"""
+        """
+        根据标签查询餐馆
+        127.0.0.1:8000/tag_restaurant/c小吃?res_name=粉
+        """
 
         # 解析参数
         param1, param2 = None, None
@@ -78,6 +127,10 @@ class TagRestaurantDetailView(GenericAPIView):
             restaurant = Restaurant.objects.filter(tag__tag_name=param1)
         if param2:
             restaurant = restaurant.filter(tag__tag_name=param2)
+
+        # 根据参数名进行查询
+        if 'res_name' in request.GET:
+            restaurant = restaurant.filter(res_name__contains=request.GET['res_name'])
         # 对返回结果进行处理
         restaurant = self.paginate_queryset(restaurant)  # 获取分页数据
         serializer = self.get_serializer(restaurant, many=True)
@@ -131,14 +184,6 @@ class CollectionModelViewSet(ModelViewSet):
         count = Collection.objects.filter(restaurant=pk).count()
         return Response(count)
 
-    @action(methods=['get'], detail=True)
-    def restaurant(self, request, pk):
-        """获取收藏的餐馆列表"""
-        restaurants = Restaurant.objects.filter(restaurant_collection__user_id=pk)
-        restaurants = self.paginate_queryset(restaurants)
-        serializer = RestaurantSerializer(instance=restaurants, many=True)
-        return self.get_paginated_response(serializer.data)
-
     @action(methods=['post'], detail=False)
     def collected(self, request):
         """检查是否收藏"""
@@ -150,7 +195,8 @@ class CollectionModelViewSet(ModelViewSet):
 class MenuModelViewSet(ModelViewSet):
     queryset = Menu.objects.all()
     serializer_class = MenuSerializer
-    pagination_class = MyPageNumberPagination
+
+    # pagination_class = MyPageNumberPagination
 
     def update(self, request, *args, **kwargs):
         partial = kwargs.pop('partial', False)
@@ -158,7 +204,8 @@ class MenuModelViewSet(ModelViewSet):
 
         try:
             # 删除之前的图片
-            os.remove(instance.picture.path)
+            if request.data['picture']:
+                os.remove(instance.picture.path)
         except Exception:
             pass
 
@@ -177,9 +224,8 @@ class MenuModelViewSet(ModelViewSet):
     def restaurant_menu(self, request, pk):
         """根据餐馆id获取菜单列表"""
         menu = Menu.objects.filter(restaurant=pk).order_by('-recommendations')
-        menu = self.paginate_queryset(menu)
         serializer = self.get_serializer(menu, many=True)
-        return self.get_paginated_response(serializer.data)
+        return Response(serializer.data)
 
     @action(methods=['post'], detail=False)
     def res_insert(self, request):
@@ -197,7 +243,7 @@ class MenuModelViewSet(ModelViewSet):
         为指定菜品添加一个推荐数
         post的body中携带参数menus，格式为[1,2,3,4]
         """
-        menus = json.loads(request.data['menus'])
+        menus = request.data['menus']
         count = Menu.objects.filter(id__in=menus).update(recommendations=F('recommendations') + 1)
         return Response(count)
 
@@ -209,19 +255,53 @@ class ReviewModelViewSet(ModelViewSet):
 
     @action(methods=['get'], detail=True)
     def restaurant_review(self, request, pk):
+        """根据餐馆获取评论"""
         reviews = Review.objects.filter(restaurant=pk)
         reviews = self.paginate_queryset(reviews)
         serializer = self.get_serializer(reviews, many=True)
         return self.get_paginated_response(serializer.data)
 
+    @action(methods=['get'], detail=True)
+    def user_review(self, request, pk):
+        """根据用户获取评论"""
+        reviews = Review.objects.filter(user=pk)
+        reviews = self.paginate_queryset(reviews)
+        serializer = self.get_serializer(reviews, many=True)
+        return self.get_paginated_response(serializer.data)
+
     def create(self, request, *args, **kwargs):
+        """创建评论"""
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         self.perform_create(serializer)
-        restaurant = request.data['restaurant']
-        score = Review.objects.filter(restaurant=restaurant).values('restaurant').annotate(
-            avg_score=Avg('score')).values_list(
-            'avg_score')[0][0]
-        Restaurant.objects.filter(id=restaurant).update(score=score)
+
+        Thread(target=self.run_update, args=(serializer,)).start()
+
         headers = self.get_success_headers(serializer.data)
         return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+
+    @staticmethod
+    def run_update(serializer):
+        restaurant = serializer.data['restaurant']
+        # 计算餐馆的平均分
+        score_res = Restaurant.objects.filter(id=restaurant).values_list('score')[0][0]
+        score = serializer.data['score']
+        count = Review.objects.filter(restaurant=restaurant).count()
+        score = ((score_res * (count - 1)) + score) / count
+        Restaurant.objects.filter(id=restaurant).update(score=score)
+        # 通过AI模型计算分数
+        AI_score = analyze(serializer.data['text'])[0]
+        Review.objects.filter(id=serializer.data['id']).update(analyze_result=AI_score)
+        # 删除词云图片
+        wordCloudPath = os.path.join(BASE_DIR, 'media', 'wordClouds', restaurant + ".png")
+        if os.path.exists(wordCloudPath):
+            os.remove(wordCloudPath)
+
+    @action(methods=['get'], detail=True)
+    def pos_or_nav_reviews(self, request, pk):
+        """获取积极或负面评论，参数传入6_0表示6号餐馆的积极评论，6_1表示6号餐馆的负面评论"""
+        res, f = pk.split('_')
+        reviews = Review.objects.filter(restaurant=res).filter(analyze_result=f)
+        reviews = self.paginate_queryset(reviews)
+        serializer = self.get_serializer(reviews, many=True)
+        return self.get_paginated_response(serializer.data)
